+++ crates/machine/Cargo.toml
@@ -14,6 +14,7 @@ colored = "2"
 substring = "1.4.5"
 env_logger = "0.10.0"
 keyboard-types = { version="0.7", features=["webdriver"] }
+log = "0.4.20"
 
 [[bin]]
 name = "machine"
