+++ crates/machine/src/machine/mos6502_machine.rs
@@ -1,7 +1,6 @@
 use super::{
     impl_reg_setter, MOS6502Memory, Machine, MachineConfig, MachineStatus, Memory, RegSetter,
 };
-use crate::debugger::DebugMachine;
 use crate::mos6502::{execute_operation, Operation, MOS6502};
 use std::num::Wrapping;
 
@@ -61,4 +60,3 @@ impl Machine for MOS6502Machine {
     }
 }
 
-impl DebugMachine for MOS6502Machine {}
